requirements.txt
PyGithub
requests
python-dotenv